// apps/client/src/app/core/swagger/model/comment.ts
/**
 * NgReports
 * NG Reports API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { UserView } from './userView';

export interface Comment { 
    commentId: string;
    author: UserView;
    content: string;
    date: Date;
}